Nagging: extract code from init.gradle.kts to an init plugin
