+++ test/internal/runtime/src/main/kotlin/net/twisterrob/gradle/nagging/agpVersion.kt
@@ -1,0 +1,4 @@
+package net.twisterrob.gradle.nagging
+
+internal val agpVersion: String = System.getProperty("net.twisterrob.test.android.pluginVersion")
+	?: error("Property 'net.twisterrob.test.android.pluginVersion' is not set.")
